+++ alicloud/import_alicloud_security_group_test.go
@@ -11,7 +11,10 @@ func TestAccAlicloudSecurityGroup_importBasic(t *testing.T) {
 	resourceName := "alicloud_security_group.foo"
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheckWithRegions(t, true, connectivity.EcsClassicSupportedRegions) },
+		PreCheck: func() {
+			testAccPreCheckWithRegions(t, true, connectivity.EcsClassicSupportedRegions)
+			testAccPreCheckWithAccountSiteType(t, DomesticSite)
+		},
 		Providers:    testAccProviders,
 		CheckDestroy: testAccCheckSecurityGroupDestroy,
 		Steps: []resource.TestStep{