+++ alicloud/resource_alicloud_instance_test.go
@@ -124,6 +124,7 @@ func TestAccAlicloudInstance_basic(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheckWithRegions(t, true, connectivity.EcsClassicSupportedRegions)
+			testAccPreCheckWithAccountSiteType(t, DomesticSite)
 		},
 
 		// module name
@@ -640,6 +641,7 @@ func TestAccAlicloudInstanceChargeType_post2Pre(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
+			testAccPreCheckWithAccountSiteType(t, DomesticSite)
 		},
 		Providers:    testAccProviders,
 		CheckDestroy: testAccCheckInstanceDestroy,
@@ -699,6 +701,7 @@ func TestAccAlicloudInstanceChargeType_pre2Post(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
+			testAccPreCheckWithAccountSiteType(t, DomesticSite)
 		},
 		Providers:    testAccProviders,
 		CheckDestroy: testAccCheckInstanceDestroy,