+++ CHANGELOG.md
@@ -6,7 +6,8 @@ FEATURES:
 
 IMPROVEMENTS:
 
-- Add account site type setting to skip unsupported test cases automatically [GH-788]
+- Improve instance and security group testcase for different account site [GH-791]
+- Add account site type setting to skip unsupported test cases automatically [GH-790]
 - Modify useragent to meet the standard of sdk [GH-778]
 - Modify kms client to dock with the alicloud official GO SDK [GH-763]